# Conflicts:
#	database/src/main/java/com/ut/database/dao/LockGroupDao.java
#	module_lock/src/main/java/com/ut/module_lock/fragment/LockListFragment.java

+++ database/src/main/java/com/ut/database/entity/LockGroup.java
@@ -12,17 +12,17 @@ public class LockGroup {
      * userId : 11
      */
     @PrimaryKey
-    private int id;
+    private long id;
     private String name;
     private long createTime;
-    private int userId;
+    private long userId;
     private int current = 0;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -42,11 +42,11 @@ public class LockGroup {
         this.createTime = createTime;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
